Add files via upload
Project based on the London Bike Share data accessible from Kaggle.com. This project was guided by the YouTube creator named "Data with Mo". The data was manipulated with Python code but the code was provided and explained by the YouTube creator. The report and visualization was created in Tableau Public.